# Conflicts:
#	src/main/java/com/example/kinobackend/movie/Movie.java
#	src/main/java/com/example/kinobackend/movie/MovieController.java
#	src/main/java/com/example/kinobackend/movie/MovieService.java
#	src/main/java/com/example/kinobackend/showing/Showing.java

+++ src/main/java/com/example/kinobackend/ticket/Ticket.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class Ticket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long ticketId;
+    private int ticketId;
 
     @OneToOne
     @JoinColumn(name = "seat_id", nullable = false)
